+++ src/main/menus/edit.js
@@ -66,5 +66,23 @@ export default {
     click: (menuItem, browserWindow) => {
       actions.edit(browserWindow, 'aidou')
     }
+  }, {
+    label: 'Insert Image',
+    submenu: [{
+      label: 'Absolute Path',
+      click (menuItem, browserWindow) {
+        actions.insertImage(browserWindow, 'absolute')
+      }
+    }, {
+      label: 'Relative Path',
+      click (menuItem, browserWindow) {
+        actions.insertImage(browserWindow, 'relative')
+      }
+    }, {
+      label: 'Upload to Cloud (EXP)',
+      click (menuItem, browserWindow) {
+        actions.insertImage(browserWindow, 'upload')
+      }
+    }]
   }]
 }